Move all of deploy's api functionality into bases.
